Update README.md
@cbugk  While I appreciate your perspective, I believe providing detailed instructions with code snippets enhances user understanding and usability, particularly for newcomers. However, I understand your concerns about the potential difficulty of reproducing these snippets. So made it that you can copy process together.

Also if you copy/paste the whole code block as once it throws an error indicating there is no ```out``` folder. I suppose it is because the program does not automatically create an out folder because of ease of use but still this information should also be provided in some kind of a documentation solution. We should consider including information about this aspect in our documentation to ensure users understand the expected behavior. This way, we can maintain clarity while also making the repository more accessible to all users.